# Conflicts:
#	analysis.cli/src/main/java/org/iobserve/analysis/service/AnalysisMain.java
#	cli.analysis/src/main/java/org/iobserve/service/cli/AnalysisMain.java

+++ monitoring/src/main/java/org/iobserve/monitoring/probe/servlet/SessionAndTraceRegistrationFilterForJPetstoreAttributeCoding.java
@@ -372,7 +372,7 @@ public class SessionAndTraceRegistrationFilterForJPetstoreAttributeCoding implem
                     final Long code = SessionAndTraceRegistrationFilterForJPetstoreAttributeCoding.codes
                             .containsKey(queryParameterSplit[1])
                                     ? SessionAndTraceRegistrationFilterForJPetstoreAttributeCoding.codes
-                                            .get(queryParameterSplit[1])
+                                            .get(queryParameterSplit[1]).longValue()
                                     : -100L;
 
                     callInformations.add(new CallInformation(queryParameterSplit[0], code));
